pyproject.toml
[project]
name = "studypython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiosignal==1.4.0",
    "alembic==1.18.4",
    "altgraph==0.17.5",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.91.0",
    "anyio==4.13.0",
    "apscheduler==3.11.2",
    "asyncpg==0.31.0",
    "attrs==26.1.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.3",
    "bilibili-api-python==17.4.1",
    "brotli==1.2.0",
    "build==1.4.2",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "chromadb==1.5.7",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.6",
    "dataclasses-json==0.6.7",
    "deepeval==3.9.6",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docx2txt==0.9",
    "durationpy==0.10",
    "ecdsa==0.19.2",
    "email-validator==2.3.0",
    "emoji==2.15.0",
    "et-xmlfile==2.0.0",
    "fastapi==0.135.2",
    "filelock==3.25.2",
    "filetype==1.2.0",
    "flashrank==0.2.10",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "google-search-results==2.4.2",
    "googleapis-common-protos==1.74.0",
    "greenlet==3.3.2",
    "grpcio==1.80.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.9.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==35.0.0",
    "langchain==1.2.15",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.26",
    "langchain-deepseek==1.0.1",
    "langchain-huggingface==1.2.1",
    "langchain-milvus==0.3.3",
    "langchain-openai==1.1.12",
    "langchain-text-splitters==1.1.1",
    "langdetect==1.0.9",
    "langfuse==4.2.0",
    "langgraph==1.1.6",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.9",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.25",
    "lark==1.3.1",
    "learning-rag-app",
    "llvmlite==0.47.0",
    "lxml==6.0.2",
    "mako==1.3.10",
    "markdown==3.10.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mlflow==2.22.4",
    "mmh3==5.2.1",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "networkx==3.6.1",
    "nltk==3.9.4",
    "numba==0.65.0",
    "numpy==2.4.4",
    "oauthlib==3.3.1",
    "olefile==0.47",
    "onnxruntime==1.24.4",
    "openai==2.30.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.41.0",
    "opentelemetry-exporter-otlp-proto-common==1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.41.0",
    "opentelemetry-exporter-otlp-proto-http==1.41.0",
    "opentelemetry-proto==1.41.0",
    "opentelemetry-sdk==1.41.0",
    "opentelemetry-semantic-conventions==0.62b0",
    "orjson==3.11.8",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "pandas==3.0.2",
    "passlib==1.7.4",
    "pefile==2024.8.26",
    "pillow==12.2.0",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "protobuf==6.33.6",
    "psutil==7.2.2",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.3",
    "pybase64==1.4.3",
    "pycparser==3.0",
    "pycryptodomex==3.23.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pygame-ce==2.5.6",
    "pygments==2.20.0",
    "pyinstaller==6.19.0",
    "pyinstaller-hooks-contrib==2026.1",
    "pyjwt==2.12.1",
    "pymilvus==2.6.12",
    "pymilvus-model==0.3.2",
    "pypdf==6.9.2",
    "pypdfium2==5.7.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-iso639==2026.1.31",
    "python-jose==3.5.0",
    "python-magic==0.4.27",
    "python-multipart==0.0.22",
    "python-oxmsg==0.0.2",
    "pywin32-ctypes==0.2.3",
    "pyyaml==6.0.3",
    "qrcode==8.2",
    "qrcode-terminal==0.8",
    "rank-bm25==0.2.2",
    "rapidfuzz==3.14.5",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.33.0",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "ruff==0.15.9",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==5.3.0",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "sqlalchemy==2.0.49",
    "sqlmodel==0.0.38",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tavily-python==0.7.23",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "torch==2.11.0",
    "tqdm==4.67.3",
    "transformers==5.5.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2026.1",
    "tzlocal==5.3.1",
    "unstructured==0.18.32",
    "unstructured-client==0.43.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uv==0.11.6",
    "uvicorn==0.42.0",
    "watchfiles==1.1.1",
    "webencodings==0.5.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "wrapt==1.17.3",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]

[tool.uv.sources]
learning-rag-app = { path = "../../python/agentproject", editable = true }
